+++ src/plonemeeting/portal/core/content/publication.py
@@ -1,7 +1,11 @@
 from AccessControl import ClassSecurityInfo
+from DateTime import DateTime
+from Products.CMFCore.permissions import ManagePortal
+from Products.CMFCore.permissions import ModifyPortalContent
+from Products.CMFCore.permissions import View
+from Products.CMFCore.utils import _checkPermission
 from collective.timestamp.behaviors.timestamp import ITimestampableDocument
 from collective.timestamp.interfaces import ITimeStamper
-from DateTime import DateTime
 from imio.helpers import EMPTY_DATETIME
 from imio.helpers.content import object_values
 from plone import api
@@ -10,6 +14,7 @@ from plone.app.contenttypes.interfaces import IFile
 from plone.app.dexterity.textindexer import searchable
 from plone.app.textfield import RichText
 from plone.app.z3cform.widgets.richtext import RichTextFieldWidget
+from plone.autoform.directives import order_after
 from plone.autoform.directives import read_permission
 from plone.autoform.directives import widget
 from plone.autoform.directives import write_permission
@@ -18,10 +23,6 @@ from plone.indexer.decorator import indexer
 from plone.namedfile.field import NamedBlobFile
 from plone.supermodel import model
 from plonemeeting.portal.core import _
-from Products.CMFCore.permissions import ManagePortal
-from Products.CMFCore.permissions import ModifyPortalContent
-from Products.CMFCore.permissions import View
-from Products.CMFCore.utils import _checkPermission
 from zope import schema
 from zope.component import getMultiAdapter
 from zope.globalrequest import getRequest
@@ -29,41 +30,37 @@ from zope.interface import implementer
 
 
 class IPublication(model.Schema, IFile, ITimestampableDocument):
-    """ Marker interface and Dexterity Python Schema for Publication
-    """
+    """Marker interface and Dexterity Python Schema for Publication"""
 
     document_type = schema.Choice(
         vocabulary="plonemeeting.portal.vocabularies.document_types",
-        title=_(u"Document type"),
+        title=_("Document type"),
         required=False,
     )
 
     category = schema.Choice(
         vocabulary="plonemeeting.portal.vocabularies.global_categories",
-        title=_(u"Category"),
+        title=_("Category"),
         required=False,
     )
 
     legislative_authority = schema.Choice(
         vocabulary="plonemeeting.portal.vocabularies.legislative_authorities",
-        title=_(u"Legislative authority"),
+        title=_("Legislative authority"),
         required=False,
     )
 
     decision_date = schema.Date(
-        title=_(u"Decision date"),
+        title=_("Decision date"),
         required=False,
     )
 
     entry_date = schema.Date(
-        title=_(u"Entry date"),
+        title=_("Entry date"),
         required=False,
     )
 
-    consultation_text = RichText(
-        title=_(u"Hour and place of consultation"),
-        required=False
-    )
+    consultation_text = RichText(title=_("Hour and place of consultation"), required=False)
 
     text = RichText(
         title=_("Text"),
@@ -72,16 +69,13 @@ class IPublication(model.Schema, IFile, ITimestampableDocument):
     widget("text", RichTextFieldWidget)
     searchable("text")
 
-    model.primary('file')
-    file = NamedBlobFile(
-        title="File",
-        accept=("application/pdf", ),
-        required=False)
+    model.primary("file")
+    file = NamedBlobFile(title="File", accept=("application/pdf",), required=False)
 
     # Styling fieldset
     model.fieldset(
         "authority",
-        label=_(u"Authority"),
+        label=_("Authority"),
         fields=[
             "subject_to_authority",
             "authority_date",
@@ -89,19 +83,19 @@ class IPublication(model.Schema, IFile, ITimestampableDocument):
         ],
     )
     subject_to_authority = schema.Bool(
-        title=_(u"Subject to authority?"),
+        title=_("Subject to authority?"),
         required=False,
         default=True,
     )
 
     authority_date = schema.Date(
-        title=_(u"Authority date"),
+        title=_("Authority date"),
         description=_("authority_date_description"),
         required=False,
     )
 
     expired_authority_date = schema.Date(
-        title=_(u"Expired authority date"),
+        title=_("Expired authority date"),
         description=_("expired_authority_date_description"),
         required=False,
     )
@@ -112,10 +106,7 @@ class IPublication(model.Schema, IFile, ITimestampableDocument):
         fields=["timestamped_file"],
     )
     write_permission(timestamped_file="cmf.ManagePortal")
-    timestamped_file = NamedBlobFile(
-        title="Timestamped file",
-        accept=("application/zip", ),
-        required=False)
+    timestamped_file = NamedBlobFile(title="Timestamped file", accept=("application/zip",), required=False)
 
     # Making sure timestamp file is accessible to anonymous.
     # By default it has a custom permission
@@ -124,8 +115,7 @@ class IPublication(model.Schema, IFile, ITimestampableDocument):
 
 @implementer(IPublication)
 class Publication(Container, File):
-    """
-    """
+    """ """
 
     security = ClassSecurityInfo()
 
@@ -141,40 +131,47 @@ class Publication(Container, File):
     def is_power_user(self):
         """ """
         institution = self._get_institution()
-        return not institution.publications_power_users or \
-            api.user.get_current().getId() in institution.publications_power_users
+        return (
+            not institution.publications_power_users
+            or api.user.get_current().getId() in institution.publications_power_users
+        )
 
     def is_timestamped(self):
         return ITimeStamper(self).is_timestamped()
 
     def may_back_to_private(self):
         """Only Manager may back to private except if
-           current review_state is "planned"."""
+        current review_state is "planned"."""
+        user = api.user.get_current()
         if api.content.get_state(self) == "planned":
-            return _checkPermission(ModifyPortalContent, self)
+            # Editor can't edit it directly, but can put it back to private to edit it.
+            return user.has_role("Editor", object=self) or _checkPermission(ManagePortal, self)
         else:
             return _checkPermission(ManagePortal, self)
 
     def may_plan(self):
         """May plan if able to modify and
-           a "publication date" (effectiveDate) is defined."""
-        return _checkPermission(ModifyPortalContent, self) and \
-            self.effective_date is not None and self.effective_date > DateTime()
+        a "publication date" (effectiveDate) is defined."""
+        return (
+            _checkPermission(ModifyPortalContent, self)
+            and self.effective_date is not None
+            and self.effective_date > DateTime()
+        )
 
     def may_publish(self):
         """May publish if able to modify and
-           a "publication date" (effectiveDate) is NOT defined.
-           When "unpublished" check that time stamp  was not modified."""
-        return _checkPermission(ModifyPortalContent, self) and \
-            (self.effective_date is None or
-             (api.content.get_state(self) == "planned" and
-              self.effective_date < DateTime()) or
-             (api.content.get_state(self) == "unpublished" and
-              self.is_power_user() and
-              self.timestamp_still_valid()) or
-             (api.content.get_state(self) == "private" and
-              self.effective_date < DateTime() and
-              self.enable_timestamping is False))
+        a "publication date" (effectiveDate) is NOT defined.
+        When "unpublished" check that time stamp  was not modified."""
+        return _checkPermission(ModifyPortalContent, self) and (
+            self.effective_date is None
+            or (api.content.get_state(self) == "planned" and self.effective_date < DateTime())
+            or (api.content.get_state(self) == "unpublished" and self.is_power_user() and self.timestamp_still_valid())
+            or (
+                api.content.get_state(self) == "private"
+                and self.effective_date < DateTime()
+                and self.enable_timestamping is False
+            )
+        )
 
     def timestamp_still_valid(self):
         """Check if timestamp still corresponds to effective_date."""
@@ -189,8 +186,8 @@ def get_decision_date(obj):
 @indexer(IPublication)
 def get_effective_date(obj):
     """As elements are sorted in the faceted on effective date, if no date, it
-       it not returned so set a date of 01/01/1950 for publications without an
-       effective date."""
+    it not returned so set a date of 01/01/1950 for publications without an
+    effective date."""
     return obj.effective_date or EMPTY_DATETIME
 
 
@@ -201,17 +198,13 @@ def get_effective_year(obj):
 
 @indexer(IPublication)
 def get_pretty_category(obj):
-    global_categories = api.portal.get_registry_record(
-        name="plonemeeting.portal.core.global_categories"
-    )
+    global_categories = api.portal.get_registry_record(name="plonemeeting.portal.core.global_categories")
     return global_categories.copy().get(obj.category, [])
 
 
 @indexer(IPublication)
 def get_pretty_document_type(obj):
-    document_types = api.portal.get_registry_record(
-        name="plonemeeting.portal.core.document_types"
-    )
+    document_types = api.portal.get_registry_record(name="plonemeeting.portal.core.document_types")
     return document_types.copy().get(obj.document_type, [])